Format and cleanup

+++ src/fx_shake.cc
@@ -1,16 +1,16 @@
 #include "fx_shake.h"
 
-Shake::Shake(const f32 magnitude, const i32 milliseconds)
+Shake::Shake(const int magnitude, const int milliseconds)
 	: FxBase(milliseconds), magnitude_(magnitude), offset_({ 0, 0 })
 {
 }
 
-void Shake::update(const f32 delta)
+void Shake::update(const float delta)
 {
 	elapsed_ += delta;
-	const f32 mag = magnitude_;
+	const int mag = magnitude_;
 	this->offset_ = Vector2{
-		(f32)GetRandomValue(-mag, mag),
-		(f32)GetRandomValue(-mag, mag)
+		(float)GetRandomValue(-mag, mag),
+		(float)GetRandomValue(-mag, mag)
 	};
 }
